make Histogram Utils and move non struct stuff from Struct headers
